requirements.txt
Django==2.2
djangorestframework==3.11.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
